fix: 月均收入份数改为工程级整数,不再除以月份
根因:用户明确要求份数是工程级,只有金额才除以月份数

1. 后端 monthly_avg_count 改为直接返回 yearSettledCount(整数),不再/monthDivisor

2. 前端 DTO monthlyAvgCount 类型 Double->Int

3. SettlementSummary.monthlyAvgCount 类型 Double->Int

4. Screen 显示逻辑简化为 toString() 直接显示整数

之前份数显示0.7是因为 yearSettledCount=8÷12月=0.67 被格式化

金额保持不变: yearSettledUserAmount/monthDivisor

+++ salary-android/core/network/src/main/java/com/salary/core/network/api/StatisticsApi.kt
@@ -69,9 +69,9 @@ data class DashboardStatsDto(
     /** 卡片3：今年创建的所有工程总额（工程级 total_amount 合计） */
     @SerialName("year_project_amount")
     val yearProjectAmount: Double = 0.0,
-    /** 卡片4：月均份数（今年已结算工程数 / 当前月份，工程级） */
+    /** 卡片4：月均份数（今年已结算工程总数，工程级整数） */
     @SerialName("monthly_avg_count")
-    val monthlyAvgCount: Double = 0.0,
+    val monthlyAvgCount: Int = 0,
     /** 卡片4：月均金额（今年个人已结算工资 / 当前月份，个人级） */
     @SerialName("monthly_avg_amount")
     val monthlyAvgAmount: Double = 0.0
